fix: вынес ssr сервера в отдельный файл
