+++ packages/client/ssr/ssr.tsx
@@ -2,8 +2,7 @@ import {App} from 'client/src/App';
 import {renderToString} from 'react-dom/server';
 import {TSsrRenderProps} from './typing';
 
-export const render: TSsrRenderProps = (url: string, data: unknown) => {
-	console.log('data for redux:', data);
+export const render: TSsrRenderProps = (url) => {
 	return renderToString(
 		<App/>,
 	);
