+++ packages/client/ssr/typing.ts
@@ -1,4 +1,3 @@
 export type TSsrRenderProps = (
 	url: string,
-	data: unknown,
 ) => string;
