+++ packages/server/server.ts
@@ -2,29 +2,25 @@ import type {ViteDevServer} from 'vite';
 import {createServer as createViteServer} from 'vite';
 import cors from 'cors';
 import express from 'express';
-import fs from 'fs';
 import path from 'path';
 
-import type {TSsrRenderProps} from 'client/ssr/typing';
+import {getSsrPath, ssrContent} from './ssr';
 
 export async function startServer(isDev: boolean, port: number) {
 	const app = express();
 	app.use(cors());
 
 	let vite: ViteDevServer;
-	const ssrDevPath = path.dirname(require.resolve('client'));
-	const ssrProdPath = require.resolve('client/ssr-dist/client.cjs');
-	let distPath: string;
 
 	if (isDev) {
 		vite = await createViteServer({
 			server: {middlewareMode: true},
-			root: ssrDevPath,
+			root: getSsrPath(isDev),
 			appType: 'custom',
 		});
 		app.use(vite.middlewares);
 	} else {
-		distPath = path.dirname(require.resolve('client/dist/index.html'));
+		const distPath = path.dirname(require.resolve('client/dist/index.html'));
 		app.use('/assets', express.static(path.resolve(distPath, 'assets')));
 	}
 
@@ -33,30 +29,9 @@ export async function startServer(isDev: boolean, port: number) {
 	});
 
 	app.use('*', async (req, res, next) => {
-		const url = req.originalUrl;
-
 		try {
-			let template: string;
-			let render: TSsrRenderProps;
-
-			if (!isDev && distPath) {
-				template = fs.readFileSync(
-					path.resolve(distPath, 'index.html'),
-					'utf-8',
-				);
-				render = (await import(ssrProdPath)).render;
-			} else {
-				template = fs.readFileSync(
-					path.resolve(ssrDevPath, 'index.html'),
-					'utf-8',
-				);
-				template = await vite.transformIndexHtml(url, template);
-				render = (await vite.ssrLoadModule(path.resolve(ssrDevPath, 'ssr/ssr.tsx'))).render;
-			}
-
-			const data = ['da ta for redux'];
-			const appHtml = render(url, data);
-			const html = template.replace('<!--ssr-outlet-->', appHtml);
+			const url = req.originalUrl;
+			const html = await ssrContent(vite, url, isDev);
 			res.status(200).set({'Content-Type': 'text/html'}).end(html);
 		} catch (e) {
 			if (isDev) {
